# openapi.yaml
openapi: 3.0.0
info:
  title: Your API
  version: 1.0.0
  description: API documentation for signup route

paths:
  /auth/signup:
    post:
      summary: Create a new user account
      description: Endpoint to register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        "201":
          description: User account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /auth/verifyemail:
    post:
      summary: Verify user email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                redirectUri:
                  type: string
                  description: The URI to redirect after verification
                email:
                  type: string
                  format: email
                  description: The user's email address
              required:
                - redirectUri
                - email
      responses:
        "200":
          description: User email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userToken:
                    type: string
                    description: Token for the verified user
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message

  /auth/deleteuser:
    delete:
      summary: Delete user account
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Deletion confirmation message
        "401":
          description: Unauthorized - Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
  /profile:
    get:
      summary: Get user profile information
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userprofile:
                    type: object # Adjust the user profile schema as needed
                    description: User profile data

  /profile/update:
    post:
      summary: Update user profile information
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object # Adjust the request body schema as needed
              properties:
                school:
                  type: string
                  description: School information
                faculty:
                  type: string
                  description: Faculty information
                department:
                  type: string
                  description: Department information
                level:
                  type: string
                  description: Level information
      responses:
        "200":
          description: User profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Update confirmation message